fixes to cpu stable diffusion

+++ ldm/modules/attention.py
@@ -251,7 +251,7 @@ class SpatialTransformer(nn.Module):
         # note: if no context is given, cross-attention defaults to self-attention
         b, c, h, w = x.shape
         x_in = x
-        x = self.norm(x)
+        x = self.norm(x.float())
         x = self.proj_in(x)
         x = rearrange(x, 'b c h w -> b (h w) c')
         for block in self.transformer_blocks:
